// src/app/api/admin/model/flow-meta.ts
/**
 * Kerberos administration API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FlowMetaData } from './flow-meta-data';


export interface FlowMeta { 
    /**
     * The name of the flow component, e.g. \"observability\", \"router\".
     */
    name: string;
    data: FlowMetaData;
}