// src/app/api/admin/model/flow-meta-data-router.ts
/**
 * Kerberos administration API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FlowMetaDataRouterBackend } from './flow-meta-data-router-backend';


export interface FlowMetaDataRouter { 
    backends?: Array<FlowMetaDataRouterBackend>;
}