// src/app/api/admin/model/flow-meta-data-router-backend.ts
/**
 * Kerberos administration API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FlowMetaDataRouterBackend { 
    name: string;
    host: string;
    port: number;
}